Improved tests
